Split CouplingCellWithLinkedCells from CouplingCell
